v0.3.91: Polish - clippy fixes and targeted lint allows
- Applied cargo clippy --fix (44 files, 29 LOC net reduction)
- Added targeted #![allow] for structural patterns:
  - clippy::too_many_arguments (detector configs)
  - clippy::if_same_then_else (pattern matching)
  - clippy::only_used_in_recursion
  - clippy::needless_update
  - clippy::upper_case_acronyms
  - clippy::type_complexity
- Fixed manual_clamp → .clamp()
- Updated DETECTORS.md: 95+ → 108 detectors
- Warnings: 87 → 6 (minor edge cases remaining)
- 21 tests pass, build clean

+++ repotoire-cli/src/detectors/ai_missing_tests.rs
@@ -498,7 +498,7 @@ impl Detector for AIMissingTestsDetector {
             
             // Get complexity from context or graph
             let complexity = if let Some(ctx) = contexts.get(&func.qualified_name) {
-                ctx.complexity.unwrap_or(1) as i64
+                ctx.complexity.unwrap_or(1)
             } else {
                 func.complexity().unwrap_or(1)
             };
